requirements.txt
ace_tools==0.0
attrs==25.1.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
emoji==0.6.0
filelock==3.17.0
fonttools==4.55.8
fsspec==2025.2.0
h11==0.14.0
huggingface==0.0.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
progressbar2==4.5.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
seaborn==0.13.2
selenium==4.29.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
